Working Ammo Changer

+++ AssaultCubeConsole/Process.h
@@ -9,7 +9,7 @@ class Process
 {
 public:
 	static DWORD GetProcessID(std::wstring processName);
-	static uint32_t GetBaseAddress(DWORD processId, std::wstring moduleName);
-	static uintptr_t CalculatePointerAddress(HANDLE handle, uint32_t ptr, std::vector<uint32_t> offsets);
+	static uintptr_t GetBaseAddress(DWORD processId, std::wstring moduleName);
+	static uintptr_t CalculatePointerAddress(HANDLE handle, uintptr_t ptr, std::vector<uintptr_t> offsets);
 };
